+++ esw-integration-test/src/test/scala/esw/integration/IntegrationTestWithAuth.scala
@@ -71,7 +71,7 @@ class IntegrationTestWithAuth extends EswTestKit(AAS) with GatewaySetup with Age
     spawnAgent(AgentSettings(agentPrefix, channel, versionConfPath))
     agentClient = AgentClient.make(agentPrefix, locationServiceUtil).rightValue
     // agent service setup
-    agentServiceWiring = AgentServiceApp.start(startLogging = false)
+    agentServiceWiring = AgentServiceApp.start(None, startLogging = false)
     val httpLocation = resolveHTTPLocation(agentServiceWiring.prefix, ComponentType.Service)
     agentService = AgentServiceClientFactory(httpLocation, () => tokenWithEswUserRole())
   }