+++ esw-sm/esw-sm-app/src/main/scala/esw/sm/app/SequenceManagerAppCommand.scala
@@ -21,6 +21,12 @@ object SequenceManagerAppCommand {
 
   @CommandName("start")
   final case class StartCommand(
+      @ExtraName("p")
+      @HelpMessage(
+        "optional argument: port on which HTTP server will be bound. " +
+          "If a value is not provided, it will be randomly picked."
+      )
+      port: Option[Int],
       @ExtraName("o")
       @HelpMessage(
         "Config file path which has mapping of sequencers and resources needed for different observing modes. This is an Optional argument which need not be provided for simulation mode"